+++ crates/ry-checker/src/infer/mod.rs
@@ -14,7 +14,7 @@ pub(crate) mod custom_operator;
 pub(crate) mod index;
 mod narrow;
 pub(crate) mod pipe;
-mod quoting;
+pub(crate) mod quoting;
 pub(crate) mod recall;
 mod switch;
 mod types;